HDDS-2444. Remove server side dependencies from ozonefs jar files
Revert "HDDS-2427. Exclude webapps from hadoop-ozone-filesystem-lib-current uber jar"

This reverts commit 6a450fbf59ef8f9c5005eb856ed9a4c38b5bcc10.
